return on dev

+++ src/Events/BotEvents/readyEvents/timedFiles/guilds.ts
@@ -1,6 +1,8 @@
 import client from '../../../../BaseClient/Bot/Client.js';
 
 export default async () => {
+ if (client.user?.id !== process.env.mainId) return;
+
  await client.util.DataBase.guilds.deleteMany({ where: { fetchat: { lt: Date.now() - 60000 } } });
 
  client.guilds.cache.forEach((g) => {
